move argument creation to own file
